refactor: restructure lib folder
