+++ src/actions/index.ts
@@ -2,4 +2,4 @@ export { signIn } from './sign-in'
 export { signOut } from './sign-out'
 export { saveContact as saveWebsite } from './common/saveContact'
 export { scheduleEvent } from './google/events/scheduler'
-export { syncGoogleCalendar } from '@/lib/data/events'
+export { syncGoogleCalendar } from '@/lib/data/google/events'
